Issue 23: Add support for installing development libraries
Added a parameter to the openssl class called "install_devel" that
defaults to false.  If it is true, then it installs openssl-devel on
Redhat systems or openssl-dev on Debian boxes.  Also explicitly fails
for operating systems not in one of those two families, or when the
install_devel parameter is not a boolean.