Fix failure to load inifile causing Puppet agent to fail.
If the inifile gem isn't installed on the node we get the following
error from Puppet agent:

> Error: Could not autoload puppet/provider/x509_cert/openssl: cannot load
> such file -- inifile
> Error: Could not autoload puppet/type/x509_cert: Could not autoload
> puppet/provider/x509_cert/openssl: cannot load such file -- inifile
> Error: Failed to apply catalog: Could not autoload puppet/type/x509_cert:
> Could not autoload puppet/provider/x509_cert/openssl: cannot load such
> file -- inifile

This happens so early in the process that we can't use Puppet to install
the required gem.

This patch changes the code so that the inifile gem isn't loaded until it
is actually needed. This has two benefits:

1. If the code that requires the inifile gem isn't actually used in this
   deployment, we don't need to install it.

2. If we need the inifile gem, we can install it through Puppet with the
   `puppet_gem` provider.