readd dependencies to class to generate configs
That was a bit fast to remove...

this reverts ae510aabcd72c60080c698687f8cbaa5402c98ec
from pull request #116 (remove dependencies since we have autorequire)
and adds the dependencies between openssl config file and cert generations.

the autorequire mentioned in the the original commit, meant the
newly available autorequire in the defined types. But that one is
an autorequire for the directory where we put the file to generate, not
the configuration file !