Use .generate to generate a new private key
In OpenSSL the private key is immutable, but we don't need that anyway.
It was previously using an indirect API to achieve the desired result
while this uses a direct method.

It also uses a let block to calculate it lazily. This means it's no
longer calculated while gathering tests but rather during the test
itself.