Support OpenSSL 3
In OpenSSL 3 the OpenSSL::PKey::EC class has become immutable.  In a
pure OpenSSL 3 world the right thing would be to use
OpenSSL::PKey.generate_key but that's not supported in OpenSSL 1.
Calling .generate is compatible with both 1 and 3.

It also stops mocking the actual calls to see if they will work.
prime239v1 is replaced by prime256v1 because on Fedora 39+ only the
latter is available.